+++ src/components/Layout2Components/NavBar2.tsx
@@ -2,6 +2,19 @@ import React from "react"
 import { useSelector, useDispatch } from "react-redux"
 import { Link, navigate } from "gatsby"
 import { Skeleton } from "@material-ui/lab"
+
+import Avatar from "@material-ui/core/Avatar"
+import Menu from "@material-ui/core/Menu"
+import MenuItem from "@material-ui/core/MenuItem"
+import MuiButton from "@material-ui/core/Button"
+import Divider from "@material-ui/core/Divider"
+import Fade from "@material-ui/core/Fade"
+import Box from "@material-ui/core/Box"
+import Tooltip from "@material-ui/core/Tooltip"
+import InputBase from "@material-ui/core/InputBase"
+import NativeSelect from "@material-ui/core/NativeSelect"
+import { makeStyles, withStyles } from "@material-ui/core/styles"
+
 import styled, { css } from "styled-components"
 
 import { rhythm } from "utils/typography"
@@ -9,7 +22,15 @@ import { navbarZ } from "consts"
 import { ThemeSlider } from "components/ThemeSlider"
 import { Button } from "components/Button"
 
-import { setPostsVersion, setCognito, RootState } from "_reduxState"
+import { Sun, Moon } from "icons"
+
+import {
+  setIsDarkMode,
+  setPostsVersion,
+  setLayoutVersion,
+  setCognito,
+  RootState,
+} from "_reduxState"
 import { useVerifyTokenSet } from "utils"
 
 const flexRowAlignItemsCenter = css`
@@ -35,71 +56,238 @@ const BarItem = styled.div`
   ${flexRowAlignItemsCenter};
 `
 
+const BootstrapInput = withStyles((theme) => ({
+  input: {
+    borderRadius: 4,
+    color: "var(--geist-foreground)",
+    position: "relative",
+    backgroundColor: "var(--geist-background)",
+    border: "1px solid var(--accents-2)",
+    padding: "0 var(--geist-space-8x)",
+    // Use the system font instead of the default Roboto font.
+    fontFamily: "var(--font-sans)",
+    fontSize: "14px",
+    height: "var(--geist-space-8x)",
+    "&:focus": {
+      borderRadius: 4,
+      boxShadow: "0 0 0 1px var(--geist-selection)",
+    },
+  },
+}))(InputBase)
+
+const menuItemLinkActiveStyle = {
+  color: "var(--geist-foreground)",
+  background: "var(--accents-1)",
+}
+
+const useStyles = makeStyles({
+  paper: {
+    background: "var(--geist-background)",
+    boxShadow: "var(--shadow-medium)",
+    color: "var(--geist-foreground)",
+  },
+  menuItemRoot: {
+    fontFamily: "var(--font-sans)",
+    fontSize: "14px",
+    color: "var(--accents-5)",
+    boxShadow: "none",
+    "&:hover": menuItemLinkActiveStyle,
+  },
+  avatarRoot: {
+    width: 38,
+    height: 38,
+    border: `2px solid var(--accents-2)`,
+  },
+  divider: {
+    background: "var(--accents-2)",
+  },
+  tooltip: {
+    backgroundColor: "var(--accents-2)",
+    color: "var(--accents-7)",
+    fontFamily: "var(--font-sans)",
+  },
+  arrow: {
+    color: "var(--accents-2)",
+  },
+  nativeSelectIcon: {
+    color: "var(--geist-foreground)",
+  },
+})
+
 /**
- * NavBar
+ * NavBar2
  * Subscribed to a few redux state changes.
  * Also dispatches actions to update the store.
  */
 const NavBar2 = () => {
-  const postsVersion = useSelector((state: RootState) => state.postsVersion)
+  const { postsVersion, layoutVersion, isDarkMode } = useSelector(
+    (state: RootState) => ({
+      postsVersion: state.postsVersion,
+      layoutVersion: state.layoutVersion,
+      isDarkMode: state.isDarkMode,
+    })
+  )
   const dispatch = useDispatch()
   const { isLoggedIn, decoded } = useVerifyTokenSet()
+
+  const [anchorEl, setAnchorEl] = React.useState(null)
+
+  const classes = useStyles()
+
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget)
+  }
+
+  const handleClose = () => {
+    setAnchorEl(null)
+  }
 
   return (
     <Bar>
       <BarItem>
-        <Button
-          style={{
-            fontWeight: 500,
-            height: 36,
-            padding: 0,
-            paddingLeft: rhythm(0.5),
-            paddingRight: rhythm(0.5),
-            margin: 0,
-            marginRight: rhythm(0.5),
-            textDecoration: postsVersion === 1 ? "underline" : undefined,
-          }}
-          onClick={() => dispatch(setPostsVersion(1))}
+        <Tooltip
+          classes={{ tooltip: classes.tooltip, arrow: classes.arrow }}
+          arrow
+          title={
+            <>
+              Toggle the "Posts" style version.
+              <br />
+              Only the home page <code>/</code> will be affected
+            </>
+          }
         >
-          V1
-        </Button>
-        <Button
-          style={{
-            fontWeight: 500,
-            height: 36,
-            padding: 0,
-            paddingLeft: rhythm(0.5),
-            paddingRight: rhythm(0.5),
-            margin: 0,
-            marginRight: rhythm(0.5),
-            textDecoration: postsVersion === 2 ? "underline" : undefined,
-          }}
-          onClick={() => dispatch(setPostsVersion(2))}
+          <Box mr={1}>
+            <Button
+              onClick={() => dispatch(setPostsVersion((postsVersion % 2) + 1))}
+            >
+              P{postsVersion}
+            </Button>
+          </Box>
+        </Tooltip>
+        <Tooltip
+          classes={{ tooltip: classes.tooltip, arrow: classes.arrow }}
+          arrow
+          title={<>Toggle the "Layout" version</>}
         >
-          V2
-        </Button>
-        <ThemeSlider />
+          <Box mr={1}>
+            <Button
+              onClick={() =>
+                dispatch(setLayoutVersion((layoutVersion % 2) + 1))
+              }
+            >
+              L{layoutVersion}
+            </Button>
+          </Box>
+        </Tooltip>
+        <Tooltip
+          classes={{ tooltip: classes.tooltip, arrow: classes.arrow }}
+          arrow
+          title={
+            <>
+              Toggle between <code>light</code> and <code>dark</code> mode
+            </>
+          }
+        >
+          <Box>
+            <ThemeSlider />
+          </Box>
+        </Tooltip>
       </BarItem>
       <BarItem>
+        {isLoggedIn === null && <Skeleton variant="text" width={"5ch"} />}
+        {isLoggedIn === false && (
+          <Link activeStyle={{ cursor: "not-allowed" }} to={"/auth/login"}>
+            Login
+          </Link>
+        )}
         {isLoggedIn === true && (
           <>
-            <Link to={"/app"}>App</Link>
-            <span>&nbsp;|&nbsp;</span>
-            <a
-              href="/playground"
-              onClick={async (event) => {
-                event.preventDefault()
-                dispatch(setCognito(null, null))
-                navigate(`/auth/login`)
-              }}
+            <MuiButton disableRipple onClick={handleClick}>
+              <Avatar classes={{ root: classes.avatarRoot }} />
+            </MuiButton>
+            <Menu
+              anchorEl={anchorEl}
+              keepMounted
+              open={Boolean(anchorEl)}
+              onClose={handleClose}
+              TransitionComponent={Fade}
+              classes={{ paper: classes.paper }}
             >
-              Logout
-            </a>
-            <span>&nbsp;|&nbsp;Hello, {decoded?.email}</span>
+              <MenuItem
+                onClick={handleClose}
+                classes={{ root: classes.menuItemRoot }}
+                component={Link}
+                activeStyle={menuItemLinkActiveStyle}
+                to="/"
+                disableRipple
+              >
+                Home
+              </MenuItem>
+
+              <MenuItem
+                onClick={handleClose}
+                classes={{ root: classes.menuItemRoot }}
+                component={Link}
+                activeStyle={menuItemLinkActiveStyle}
+                to="/app/profile"
+                disableRipple
+              >
+                Profile
+              </MenuItem>
+
+              <Box py={1}>
+                <Divider classes={{ root: classes.divider }} />
+              </Box>
+
+              <MenuItem disableRipple classes={{ root: classes.menuItemRoot }}>
+                <Box marginRight={1.5}>Theme:</Box>
+                <Box position="relative">
+                  <NativeSelect
+                    value={isDarkMode ? "dark" : "light"}
+                    variant="outlined"
+                    onChange={(e) =>
+                      dispatch(setIsDarkMode(e.currentTarget.value === "dark"))
+                    }
+                    input={<BootstrapInput />}
+                    classes={{ icon: classes.nativeSelectIcon }}
+                  >
+                    <option value={"dark"}>Dark</option>
+                    <option value={"light"}>Light</option>
+                  </NativeSelect>
+                  <Box
+                    alignItems="center"
+                    display="flex"
+                    position="absolute"
+                    top="0"
+                    left="5px"
+                    height="100%"
+                    style={{
+                      pointerEvents: "none",
+                    }}
+                  >
+                    {isDarkMode ? <Moon /> : <Sun />}
+                  </Box>
+                </Box>
+              </MenuItem>
+
+              <Box py={1}>
+                <Divider classes={{ root: classes.divider }} />
+              </Box>
+
+              <MenuItem
+                onClick={async () => {
+                  handleClose()
+                  dispatch(setCognito(null, null))
+                  await navigate(`/auth/login`)
+                }}
+                classes={{ root: classes.menuItemRoot }}
+                disableRipple
+              >
+                Logout
+              </MenuItem>
+            </Menu>
           </>
         )}
-        {isLoggedIn === false && <Link to={"/auth/login"}>Login</Link>}
-        {isLoggedIn === null && <Skeleton variant="text" width={"5ch"} />}
       </BarItem>
     </Bar>
   )