+++ src/components/layout2.tsx
@@ -1,7 +1,6 @@
 import React, { memo } from "react"
 import { useSelector } from "react-redux"
 import { useSpring, animated, OpaqueInterpolation } from "react-spring"
-import { compose } from "redux"
 import { Link, PageProps } from "gatsby"
 import styled from "styled-components"
 
@@ -26,7 +25,7 @@ interface Props {
   location: PageProps["location"]
   title: string
 }
-const Layout: React.FC<Props> = ({ location, title, children }) => {
+const Layout2: React.FC<Props> = ({ location, title, children }) => {
   const rootPath: string = `${__PATH_PREFIX__}/`
   const topLink =
     location.pathname === rootPath ? <h1>{title}</h1> : <h3>← 🏠</h3>
@@ -48,8 +47,13 @@ const Layout: React.FC<Props> = ({ location, title, children }) => {
         style={{
           marginLeft: `auto`,
           marginRight: `auto`,
-          maxWidth: rhythm(location.pathname === rootPath ? 36 : 24),
-          padding: `${rhythm(1.5)} ${rhythm(3 / 4)}`,
+          maxWidth:
+            location.pathname === rootPath
+              ? rhythm(36)
+              : location.pathname.startsWith("/app")
+              ? "var(--geist-page-width-with-margin)"
+              : rhythm(24),
+          padding: `${rhythm(1.5)} var(--geist-gap)`,
         }}
       >
         <Link
@@ -74,7 +78,7 @@ const Layout: React.FC<Props> = ({ location, title, children }) => {
   )
 }
 
-export default compose(memo)(Layout)
+export default memo(Layout2)
 
 const BlobHolder = styled.div`
   display: flex;
@@ -103,16 +107,40 @@ const D2 =
 const D3 =
   "M385,276.5Q341,303,348,367.5Q355,432,302.5,391Q250,350,205.5,377Q161,404,123,374.5Q85,345,102.5,297.5Q120,250,135,221Q150,192,138,114Q126,36,188,38.5Q250,41,308.5,44Q367,47,373.5,111Q380,175,404.5,212.5Q429,250,385,276.5Z"
 
-const Blob = ({ y }: BlobProps) => {
+const Blob = memo(({ y }: BlobProps) => {
   return (
     <animated.svg viewBox="0 0 500 500" width="100%">
+      {/* POC for react-spring animated SVG gradient background */}
+      {/* <animated.defs>
+        <animated.linearGradient
+          id="blobGradient"
+          gradientTransform="rotate(90)"
+        >
+          <animated.stop
+            offset="0%"
+            stopColor={y.interpolate({
+              range: [0, 0.25, 0.5, 0.75, 1],
+              output: ["#ffecde", "#ffccdd", "#ff9a9e", "#ef8a8a", "#a18cd1"],
+            })}
+          />
+          <animated.stop
+            offset="100%"
+            stopColor={y.interpolate({
+              range: [0, 0.25, 0.5, 0.75, 1],
+              output: ["#ffd1ff", "#fcb69f", "#fcb3ef", "#fda085", "#fbc2eb"],
+            })}
+          />
+        </animated.linearGradient>
+      </animated.defs> */}
+
       <animated.path
         d={y.interpolate({ range: [0, 0.5, 1], output: [D1, D2, D3] })}
         fill={y.interpolate({
           range: [0, 0.33, 0.66, 1],
           output: [Colors.VIOLET, Colors.ALERT, Colors.PURPLE, Colors.CYAN],
         })}
+        // fill="url('#blobGradient')"
       ></animated.path>
     </animated.svg>
   )
-}
+})