+++ src/icons/index.ts
@@ -1,2 +1,4 @@
 export { FacebookIcon } from "./facebook-icon"
 export { QuestionCircle } from "./question-circle"
+export { Moon } from "./moon"
+export { Sun } from "./sun"